pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "prism"
version = "0.1.0"
description = "Performance Reporting & Insight for Standardized Monitoring"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "pyyaml>=6.0",
    "click>=8.1.0",
    "pandas>=2.0.0",
    "plotly>=5.14.0",
    "tabulate>=0.9.0",
]

[project.optional-dependencies]
cap = [
    "cap",
]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
prism = "prism.cli:cli"

[tool.setuptools.packages.find]
include = ["prism*"]

[tool.setuptools.package-data]
prism = ["templates/**/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
